% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calcACF}
\alias{calcACF}
\title{ACF}
\usage{
calcACF(samObject, ref, sizeStart = 15, sizeEnd = 30, regIni = 1,
  regEnd = 100)
}
\arguments{
\item{samObject}{XXXX}

\item{ref}{xxx}

\item{sizeStart}{xxx}

\item{sizeEnd}{xxx}

\item{regIni}{xxx}

\item{regEnd}{xxx}
}
\description{
calculate ACF
}
