% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calcDensityPerBase}
\alias{calcDensityPerBase}
\title{Calculate}
\usage{
calcDensityPerBase(samObject, ref, readSize = 0, bin = 1, refSize = NULL,
  norm = NULL, xmin = NULL, xmax = NULL, ymax = NULL, ymin = NULL)
}
\arguments{
\item{samObject}{XXXX}

\item{ref}{xxx}

\item{readSize}{xxx}

\item{bin}{xxx}

\item{refSize}{xxx}

\item{norm}{xxx}

\item{xmin}{xxx}

\item{xmax}{xxx}

\item{ymax}{xxx}

\item{ymin}{xxx}
}
\description{
calculating Density Per Base
}
