% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calcPhasing}
\alias{calcPhasing}
\title{Calculate phasing value}
\usage{
calcPhasing(samObject, ref, extend_upstream = 20, extend_downstream = 20,
  readSize = 0, bin = 1, refSize = NULL, norm = NULL,
  region_start = NULL, region_end = NULL, ymax = NULL, ymin = NULL)
}
\arguments{
\item{samObject}{XXXX}

\item{ref}{xxx}

\item{extend_upstream}{xxx}

\item{extend_downstream}{xxx}

\item{readSize}{xxx}

\item{bin}{xxx}

\item{refSize}{xxx}

\item{norm}{xxx}

\item{region_start}{xxx}

\item{region_end}{xxx}

\item{ymax}{xxx}

\item{ymin}{xxx}
}
\description{
calculate phasing
}
