% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createWebLogo}
\alias{createWebLogo}
\title{weblogo function}
\usage{
createWebLogo(samObject, ref, sizeStart = 18, sizeEnd = 30)
}
\arguments{
\item{samObject}{XXXX}

\item{ref}{xxx}

\item{sizeStart}{xxx}

\item{sizeEnd}{xxx}
}
\description{
plot web logo
}
