% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{loadSam}
\alias{loadSam}
\title{Read and load SAM file}
\usage{
loadSam(f, sizeStart = 11, sizeEnd = 1000)
}
\arguments{
\item{f}{SAM file path.}

\item{sizeStart}{Minimum value of size to evaluate}

\item{sizeEnd}{Maximum value of size to evaluate}
}
\description{
read and load a sam file if the input is a valid path
}
