% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{loadSamLikeBed}
\alias{loadSamLikeBed}
\title{Read a SAM file and return it in BED format}
\usage{
loadSamLikeBed(f, sizeStart = 11, sizeEnd = 1000)
}
\arguments{
\item{f}{SAM file path.}

\item{sizeStart}{Minimum value of size to evaluate}

\item{sizeEnd}{Maximum value of size to evaluate}
}
\description{
read a SAM file and convert the structure to a BED file
}
