% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multiplot}
\alias{multiplot}
\title{multiplot}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{xxx}

\item{plotlist}{xxx}

\item{file}{xxx}

\item{cols}{xxx}

\item{layout}{xxx}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
plot multiple something
function created by Winston Chang,
in cookbook for R
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
