% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotSizeDistribution}
\alias{plotSizeDistribution}
\title{Size distribution plots}
\usage{
plotSizeDistribution(samObject, ref = NULL, stranded = TRUE, norm = NULL,
  sizeStart = NULL, sizeEnd = NULL, ymax = NULL, ymin = NULL)
}
\arguments{
\item{samObject}{XXXX}

\item{ref}{xxx}

\item{stranded}{xxx}

\item{norm}{xxx}

\item{sizeStart}{xxx}

\item{sizeEnd}{xxx}

\item{ymax}{xxx}

\item{ymin}{xxx}
}
\description{
This function returns a ggplot object with ....
}
