% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{loadSam}
\alias{loadSam}
\title{Read and load SAM file}
\usage{
loadSam(f, sizeStart = 11, sizeEnd = 1000)
}
\arguments{
\item{f}{SAM file path.}

\item{sizeStart}{}

\item{sizeEnd}{}
}
\description{
loadSam ...
}
