% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multiplot}
\alias{multiplot}
\title{multiplot}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{xxx}

\item{plotlist}{xxx}

\item{file}{xxx}

\item{cols}{xxx}

\item{layout}{xxx}
}
\description{
plot multiple something
}
