% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multiplot}
\alias{multiplot}
\title{multiplot}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{x}{vetor de números.}
}
\description{
Recebe um vetor de números e retorna um vetor de números somando dois
}
